Sky (illuminance) tracker

Skytrack compares geolocator based light measurements in lux with those modelled by the sky illuminance model of xyz.

skytrackr(
  data,
  start_location,
  tolerance = 15,
  iterations = 20,
  particles = 100,
  range = c(0.32, 150),
  bbox = c(-180, -90, 180, 90),
  scale = c(1, 10),
  control = list(sampler = "SMC", settings = list(initialParticles = particles,
    iterations = iterations, message = FALSE)),
  plot = FALSE,
  verbose = TRUE
)

Arguments

data

a data frame containing date time and lux values

start_location

start location of logging, required when using a tolerance based estimation. When no start_location is used the default bounding box (bbox) settings are used.

tolerance

tolerance on the search window for optimization, given in degrees (to pad left - right, top - bottom). Uses the previous steps location to constrain the parameter (location) search space. This requires the start_location parameter to be set, as you need a start position for a trusted initial location.

iterations

number of optimization iterations

particles

number of particles when using the default particle filter optimization

range

range of values to consider during processing, should be provided in lux c(min, max) or the equivalent if non-calibrated

bbox

bounding box of the location search domain given as c(xmin, ymin, xmax, ymax)

scale

scale / sky condition factor, by default covering the skylight() range of 1-10 but can be extended for more flexibility in case of non lux measurements

control

control settings for the Bayesian optimization, generally should not be altered (defaults to a sequential monte carlo method)

plot

plot map of incrementally changing determined locations as a progress method

verbose

given feedback including a progress bar

Value

data frame with location estimate, their uncertainties, and ancillary model parameters useful in quality control

Details

Model fits are applied by default to values up to sunrise or after sunset only as most critical to the model fit (capturing daylength, i.e. latitude and the location of the diurnal pattern - longitudinal displacement).