Plot seasonal profiles

Provides static or dynamic (plotly) seasonal profile plot

stk_profile(data, logger, range = c(0, 1e+05), center = "day", plotly = FALSE)

Arguments

data

A skytrackr compatible data frame.

logger

The logger to plot.

range

The light range to plot.

center

Set the data to center data on "day" or "night" (default = "day").

plotly

Logical, convert to dynamic plotly plot or not (default = FALSE)

Value

A static or dynamic graph of light levels for a given logger.